Add Python [] indexing via {} indexing of @pyobject
Octave doesn't have [] indexing.  Its possible we will later want to
use () indexing for lists, but for now {} to [] is easier to implement.
Note we also substract one from integer indexing because Python indexes
from 0 whereas Python indexes from 1.  This induces certain bugs: we
add an xtest for a bug on dicts.  In addition to integer indexing, we
also support strings, useful for string keys in dicts (tested).
We support the special case of ":" which is mapped onto the same
Python notation (tested).  Multidimensional indexing is supported and
is tested with NumPy arrays.

*pyobject/pyobject.m: implement some forms of {} indexing.