pyobject: constructor should modify the original object
Before the ctor code, Octave has made an object for us, and expects
us to populate it.  The ctor should make changes to that given object
instead of replacing it with some other `pyobject`!  In practice this
is not hard because a `pyobject` really just stores its `id`: so copy
the `id` from the other `pyobject`.

Leave a note at the top of the ctor to remind future devs.  Not sure
how to write a test for this, so I didn't.

Fixes #22.